+++ src/kernel/kmmap.c
@@ -4,6 +4,7 @@
 
 #include <stdint.h>
 #include <stddef.h>
+#include <string.h>
 
 #include <kmmap.h>
 #include <assert.h>
@@ -12,7 +13,7 @@ void kmmap_init(kmmap_t* kmmap) {
 	assert(kmmap != NULL);
 	memset(kmmap, 0, sizeof(kmmap_t));
 }
-void kmmap_add(kmmap_t* kmmap, uint64_t address, uint64_t size, uint32_t flags) {
+void kmmap_add(kmmap_t* kmmap, uintptr_t address, size_t size, uint32_t flags) {
 	assert(kmmap != NULL);
 	assert(kmmap->count < KMMAP_MAX_REGIONS);
 	kmmap->regions[kmmap->count].address = address;
@@ -20,7 +21,7 @@ void kmmap_add(kmmap_t* kmmap, uint64_t address, uint64_t size, uint32_t flags) 
 	kmmap->regions[kmmap->count].flags = flags | KMREGION_VALID;
 	kmmap->count++;
 }
-void kmmap_remove(kmmap_t* kmmap, uint64_t address) {
+void kmmap_remove(kmmap_t* kmmap, uintptr_t address) {
 	assert(kmmap != NULL);
 	for (size_t i = 0; i < kmmap->count; ++i) {
 		if (kmmap->regions[i].address == address) {
